+++ src/airflow/traits/log.rs
@@ -12,6 +12,6 @@ pub trait LogOperations: Send + Sync {
         dag_id: &str,
         dag_run_id: &str,
         task_id: &str,
-        task_try: u16,
+        task_try: u32,
     ) -> Result<Log>;
 }
